+++ .gitignore
@@ -5,6 +5,6 @@
 
 /.git
 /.idea
-.git/
+.git/*
 .idea/*
 package-lock.json
