+++ page.html
@@ -1,0 +1,23 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>my first twig</title>
+</head>
+<body>
+    
+    <header>
+        <h1>Gulp and Twig.js</h1>
+    </header>
+    <p>
+        This page is generated by Twig.js using the gulp-twig gulp plugin.
+    </p>
+    <ul>
+        
+    </ul>
+
+</body>
+</html>
