+++ src/js/app.js
@@ -1,0 +1,3 @@
+// import test from './app/index'
+
+console.log('tests');
