+++ src/js/app/index.js
@@ -1,0 +1,4 @@
+const test = () => 'first';
+
+
+export default test;
