+++ src/views/page.twig
@@ -1,0 +1,17 @@
+{% extends "template.twig" %}
+
+{% set title = 'my first twig' %}
+
+{% block page %}
+    <header>
+        <h1>Gulp and Twig.js</h1>
+    </header>
+    <p>
+        This page is generated by Twig.js using the gulp-twig gulp plugin.
+    </p>
+    <ul>
+        {% for value in benefits %}
+            <li>{{ value }}</li>
+        {% endfor %}
+    </ul>
+{% endblock %}
